+++ include/caffeine/Interpreter/Interpreter.h
@@ -138,7 +138,6 @@ private:
   ExecutionResult visitAssert(llvm::CallBase& inst);
   ExecutionResult visitSymbolicAlloca(llvm::CallBase& inst);
 
-  ExecutionResult visitMalloc(llvm::CallBase& inst);
   ExecutionResult visitCalloc(llvm::CallBase& inst);
   ExecutionResult visitFree(llvm::CallBase& inst);
   ExecutionResult visitMallocAlign(llvm::CallBase& inst);
