+++ src/Interpreter/Memory.cpp
@@ -2,10 +2,24 @@
 
 namespace caffeine {
 
-LLVMValue Interpreter::memoryAllocate(llvm::CallBase& call, const OpRef& size,
-                                      const OpRef& alignment,
-                                      const llvm::DataLayout& layout,
-                                      int address_space) {
+// Used to implement posix_memalign
+ExecutionResult Interpreter::visitMallocAlign(llvm::CallBase& call) {
+  auto size = ctx->lookup(call.getArgOperand(0)).scalar().expr();
+  auto alignment = ctx->lookup(call.getArgOperand(1)).scalar().expr();
+
+  CAFFEINE_ASSERT(call.getNumArgOperands() == 2,
+                  "Invalid malloc align signature");
+  CAFFEINE_ASSERT(call.getType()->isPointerTy(),
+                  "Invalid malloc align signature");
+
+  const llvm::DataLayout& layout = call.getModule()->getDataLayout();
+  unsigned address_space = call.getType()->getPointerAddressSpace();
+
+  CAFFEINE_ASSERT(size->type().is_int(), "Invalid malloc align signature");
+  CAFFEINE_ASSERT(size->type().bitwidth() ==
+                      layout.getIndexSizeInBits(address_space),
+                  "Invalid malloc align signature");
+
   auto ptr_width = layout.getPointerSizeInBits(address_space);
 
   if (options.malloc_can_return_null) {
@@ -22,8 +36,12 @@ LLVMValue Interpreter::memoryAllocate(llvm::CallBase& call, const OpRef& size,
       AllocOp::Create(size_op, ConstantInt::Create(llvm::APInt(8, 0xDD))),
       AllocationKind::Malloc, AllocationPermissions::ReadWrite, *ctx);
 
-  return LLVMValue(Pointer(
-      alloc, ConstantInt::Create(llvm::APInt(ptr_width, 0)), address_space));
+  ctx->stack_top().insert(
+      &call,
+      LLVMValue(Pointer(alloc, ConstantInt::Create(llvm::APInt(ptr_width, 0)),
+                        address_space)));
+
+  return ExecutionResult::Continue;
 }
 
 } // namespace caffeine
